fix:  displacement dialog-module
