+++ src/app/app.module.ts
@@ -15,7 +15,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { FooterComponent } from './shared/components/footer/footer.component';
 import { appReducer } from './shared/store/app.reducer';
 import { ModalConfirmComponent } from './shared/components/modal-confirm/modal-confirm.component';
-import { CountFiledFormService } from './pages/board-page/services/modal-prompt.cervice';
+import { ModalService } from './dialog/service/modal-prompt.service';
 import { TranslocoRootModule } from './transloco-root.module';
 import { UrlInterceptor } from './shared/interceptors/url.interceptor';
 
@@ -43,7 +43,7 @@ import { UrlInterceptor } from './shared/interceptors/url.interceptor';
     TranslocoRootModule,
   ],
   providers: [
-    CountFiledFormService,
+    ModalService,
     {
       provide: HTTP_INTERCEPTORS,
       useClass: HeaderAddTokkenInterceptor,
