+++ src/app/dialog/message-user/message-user.component.scss
@@ -21,7 +21,7 @@
   border-radius: 7px;
   text-align: center;
   color: $light-font-color;
-  background-image: url('./../../../../../../assets/img/dark-fon.jpg');
+  background-image: url('./../../../assets/img/dark-fon.jpg');
   background-position: center center;
   background-size: 100%;
   transform: translate(-50%, -50%);
