--- src/app/pages/board-page/services/modal-prompt.cervice.ts
@@ -1,19 +1,0 @@
-import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
-
-@Injectable()
-export class CountFiledFormService {
-  private oneFiledForm$ = new BehaviorSubject<boolean>(false);
-
-  public isOneFiledForm(): boolean {
-    return this.oneFiledForm$.getValue();
-  }
-
-  public setOneFiledForm() {
-    this.oneFiledForm$.next(true);
-  }
-
-  public setTwoFiledForm() {
-    this.oneFiledForm$.next(false);
-  }
-}
