+++ src/app/pages/welcome-page/components/advantage-card/advantage-card.component.html
@@ -1,6 +1,6 @@
 <div class="wrapper">
   <div class="info-box">
-    <img src="{{advantageData.iconURL}}" alt="">
+    <img src="{{advantageData.iconURL}}" alt="img">
     <p>{{advantageData.text}}</p>
   </div>
 </div>
