+++ src/app/pages/welcome-page/welcome-page.component.html
@@ -5,7 +5,7 @@
       <p class="subtitle">organizer</p>
       <button mat-raised-button>{{t('welcome.btn')}}</button>
     </div>
-    <img class="main-img" src="./assets/img/img-welcome-page.png" alt=""/>
+    <img class="main-img" src="./assets/img/img-welcome-page.png" alt="img"/>
   </div>
 </section>
 <section class="section-wrapper light-bg" *transloco="let t">
